cloud-functions/packages/ai-cs-evaluator/case-study-evaluator/requirements.txt
openai==0.28.1
pydantic==2.4.2
python-docx